Adding database and dbController.
Addition of a database and database controller. This will allow for words to be stored in a central area that can be saved and updated with words as the user wishes. Finally, the methods have been tested with unit testing. These test have been made in a way to always be consistent as it is not possible to test a variable database with retrieving the entire dataset.